Many improvements
Writes IRFinder as a single reference in gz format
IRFinder now does not require name sorted bams. Can allow coordinate or other order. As long as the paired reads are close to each other.
Added ability to read gz files
Added mappability read generation to generate specific regions (but size-wise makes no difference)

+++ src/BAMReader.h
@@ -18,8 +18,8 @@ class BAMReader {
         int IS_EOF;
         int IS_FAIL;
         
-        int IS_LENGTH;
-        int EOF_POS;
+        size_t IS_LENGTH;
+        size_t EOF_POS;
         
  		static const int bamEOFlength = 28;
 		static const char bamEOF[bamEOFlength+1];
